:recycle: Sepalate model files
