+++ src/main/kotlin/game/models/GameState.kt
@@ -1,0 +1,7 @@
+package game.models
+
+sealed class GameState {
+    object NotStart : GameState()
+    data class Playing(val turn: Turn) : GameState()
+    object Ended : GameState()
+}
