+++ src/main/kotlin/game/models/Piece.kt
@@ -1,0 +1,8 @@
+package game.models
+
+enum class Piece {
+    Empty,
+    White,
+    Black,
+    ;
+}
