+++ src/main/kotlin/game/models/Turn.kt
@@ -1,0 +1,12 @@
+package game.models
+
+enum class Turn {
+    Black,
+    White,
+    ;
+
+    fun toggle(): Turn = when (this) {
+        Black -> White
+        White -> Black
+    }
+}
